+++ validation/validator.go
@@ -27,9 +27,7 @@ import (
 
 var (
 	ErrInvalidEventType = errors.New("event type doesn't match")
-
-	errNewerBootTime  = errors.New("newer boot-time found")
-	errDuplicateEvent = errors.New("duplicate event found")
+	ErrNonEvent         = errors.New("not an event")
 )
 
 // Validator validates an event, returning false and an error if the event is not valid
@@ -112,66 +110,15 @@ func BirthdateValidator(tv TimeValidation) ValidatorFunc {
 }
 
 // DestinationValidator takes in a regex and returns a ValidatorFunc that checks if an
-// Event's destination is valid against this regex.
+// Event's destination is valid against the EventRegex and this regex.
 func DestinationValidator(regex *regexp.Regexp) ValidatorFunc {
 	return func(e interpreter.Event) (bool, error) {
+		if !interpreter.EventRegex.MatchString(e.Destination) {
+			return false, InvalidEventErr{OriginalErr: ErrNonEvent}
+		}
+
 		if !regex.MatchString(e.Destination) {
 			return false, InvalidEventErr{OriginalErr: ErrInvalidEventType}
-		}
-		return true, nil
-	}
-}
-
-// NewestBootTimeValidator returns a ValidatorFunc to check and see if an event's boot-time is
-// less than or equal to the latestEvent's boot-time. NewestBootTimeValidator assumes that latestEvent
-// has a valid boot-time and does not do any error-checking of latestEvent's boot-time.
-func NewestBootTimeValidator(latestEvent interpreter.Event) ValidatorFunc {
-	latestBootTime, _ := latestEvent.BootTime()
-	return func(e interpreter.Event) (bool, error) {
-		// event is latestEvent, no need to compare boot-times
-		if e.TransactionUUID == latestEvent.TransactionUUID {
-			return true, nil
-		}
-
-		bootTime, err := e.BootTime()
-		if err != nil || bootTime <= 0 {
-			return true, nil
-		}
-
-		// if this event has a boot-time more recent than the latest one, return an error
-		if bootTime > latestBootTime {
-			return false, InvalidEventErr{OriginalErr: InvalidBootTimeErr{OriginalErr: errNewerBootTime}}
-		}
-
-		return true, nil
-	}
-}
-
-// UniqueEventValidator returns a ValidatorFunc to check and see if compareEvent is unique, meaning that
-// it does not share the same destination type and boot-time as another event.
-// UniqueEventValidator assumes that compareEvent has a valid boot-time
-// and does not do any error-checking of compareEvent's boot-time.
-func UniqueEventValidator(compareEvent interpreter.Event, eventType *regexp.Regexp) ValidatorFunc {
-	destValidator := DestinationValidator(eventType)
-	latestBootTime, _ := compareEvent.BootTime()
-	return func(e interpreter.Event) (bool, error) {
-		// event is latestEvent, no need to compare boot-times
-		if e.TransactionUUID == compareEvent.TransactionUUID {
-			return true, nil
-		}
-
-		// see if event is the type we are looking for
-		if destMatch, _ := destValidator.Valid(e); destMatch {
-			bootTime, err := e.BootTime()
-			if err != nil || bootTime <= 0 {
-				return true, nil
-			}
-
-			// If the boot-time is the same as the latestBootTime, and the birthdate is older or equal,
-			// this means that compareEvent is a duplicate.
-			if bootTime == latestBootTime && e.Birthdate <= compareEvent.Birthdate {
-				return false, InvalidEventErr{OriginalErr: errDuplicateEvent}
-			}
 		}
 
 		return true, nil